integrate dp params in dp synthpipe
